Bug squashing
There was a bug where it certain tags and categories weren't showing up in the main list. The main list will filter out any categories/tags that have a zero post count. The tag/category models were failing to update their counts correctly, I added a portion to the post model updating it's own category and tag counts when the post is updated.